requirements.txt
attrs==23.1.0
bcrypt==4.1.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
commonmark==0.9.1
cryptography==41.0.7
future==0.18.3
gns3fy==0.8.0
idna==3.4
iniconfig==2.0.0
netmiko==4.3.0
ntc-templates==4.0.1
packaging==23.2
paramiko==2.11.1
pexpect==4.9.0
ping3==4.0.4
pluggy==1.3.0
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pydantic==1.10.13
pygments==2.17.2
PyNaCl==1.5.0
pyserial==3.5
pytest==7.1.3
PyYAML==5.3.1
requests==2.31.0
rich==12.4.4
scp==0.14.5
six==1.16.0
tabulate==0.8.10
tenacity==8.2.3
textfsm==1.1.2
tomli==2.0.1
typing-extensions==4.9.0
urllib3==2.0.5
pytest_html
